+++ package.json
@@ -19,7 +19,7 @@
     "express-validator": "^6.14.2",
     "http-status-codes": "^2.2.0",
     "method-override": "^3.0.0",
-    "mongoose": "^6.3.8",
+    "mongoose": "^6.4.1",
     "passport": "^0.6.0",
     "passport-local-mongoose": "^7.1.2"
   }
